change return type of main to int

+++ 4x4_matrix_part_1/main.c
@@ -56,7 +56,7 @@ void draw(uint8_t buffer[4][4], uint8_t delay) {
     }
 }
 
-uint8_t main(void) {    
+int main(void) {    
     
     init();
